+++ batch_scripts/linux/installation/install_mysql.sh
@@ -9,6 +9,13 @@ export SCRIPT_FOLDER="$( dirname "$(readlink -f -- "$0")" )"
 source $SCRIPT_FOLDER/set_tcapy_env_vars.sh
 
 if [ $DISTRO == "ubuntu" ]; then
+    # mkdir -p /tmp
+    # cd /tmp
+    # sudo curl -OL https://dev.mysql.com/get/mysql-apt-config_0.8.13-1_all.deb
+    # yes | sudo dpkg -i mysql-apt-config*
+    # sudo rm -f mysql-apt-config*
+    # sudo apt update
+    # sudo apt install --yes mysql-server
     sudo apt-get install --yes mysql-server
 elif [ $DISTRO == "redhat"  ]; then
     # update repo file (assumes MySQL8)
@@ -21,6 +28,8 @@ elif [ $DISTRO == "redhat"  ]; then
 
     sudo yum install --yes mysql-community-server
 fi
+
+sudo service mysql start
 
 # Will set various parameters like passwords
 sudo mysql_secure_installation